use cpp file for function
